ai-caretaker-backend/requirements.txt
fastapi
uvicorn
openai
psycopg
python-dotenv